+++ apps/backend/src/main/java/com/intranet/backend/config/AsyncConfig.java
@@ -1,0 +1,10 @@
+package com.intranet.backend.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@Configuration
+@EnableAsync
+public class AsyncConfig {
+    // Apenas a anotação @EnableAsync é necessária
+}
