+++ apps/backend/src/main/java/com/intranet/backend/config/EmailConfig.java
@@ -2,6 +2,7 @@ package com.intranet.backend.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.context.support.ResourceBundleMessageSource;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
@@ -15,11 +16,11 @@ import java.util.Collections;
 import java.util.Properties;
 
 @Configuration
-public class MailConfig {
+public class EmailConfig {
 
     // Configuração do JavaMailSender
     @Bean
-    public JavaMailSender getMailSender() {
+    public JavaMailSender mailSender() {
         JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
 
         // Estas propriedades serão substituídas pelas que estão no application.properties
@@ -36,9 +37,9 @@ public class MailConfig {
         return mailSender;
     }
 
-    // Configuração do Template Resolver para HTML
+    // Template resolver HTML para emails
     @Bean
-    public ITemplateResolver htmlTemplateResolver() {
+    public ITemplateResolver emailTemplateResolver() {
         ClassLoaderTemplateResolver templateResolver = new ClassLoaderTemplateResolver();
         templateResolver.setOrder(1);
         templateResolver.setResolvablePatterns(Collections.singleton("email/*"));
@@ -50,16 +51,38 @@ public class MailConfig {
         return templateResolver;
     }
 
-    // Configuração do Template Engine
+    // Template resolver geral
+    @Bean
+    public ITemplateResolver generalTemplateResolver() {
+        ClassLoaderTemplateResolver templateResolver = new ClassLoaderTemplateResolver();
+        templateResolver.setOrder(2);
+        templateResolver.setPrefix("templates/");
+        templateResolver.setSuffix(".html");
+        templateResolver.setTemplateMode(TemplateMode.HTML);
+        templateResolver.setCharacterEncoding("UTF-8");
+        return templateResolver;
+    }
+
+    // Template engine específico para emails
     @Bean
     public TemplateEngine emailTemplateEngine() {
         SpringTemplateEngine templateEngine = new SpringTemplateEngine();
-        templateEngine.addTemplateResolver(htmlTemplateResolver());
+        templateEngine.addTemplateResolver(emailTemplateResolver());
         templateEngine.setTemplateEngineMessageSource(emailMessageSource());
         return templateEngine;
     }
 
-    // Configuração de mensagens para templates
+    // Template engine geral
+    @Bean
+    @Primary
+    public SpringTemplateEngine templateEngine() {
+        SpringTemplateEngine templateEngine = new SpringTemplateEngine();
+        templateEngine.setTemplateResolver(generalTemplateResolver());
+        templateEngine.setTemplateEngineMessageSource(emailMessageSource());
+        return templateEngine;
+    }
+
+    // Fonte de mensagens compartilhada
     @Bean
     public ResourceBundleMessageSource emailMessageSource() {
         ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
